+++ Core.lua
@@ -103,27 +103,6 @@ enum "NameStyle" {
 GuildColor = Color(0.25, 1, 0.25, 1)
 
 function OnLoad()
-    -- make mask frame strata lower
-    local function OnInit(self, mask)
-        Style[mask]             = {
-            frameLevel          = 8,
-            frameStrata         = "DIALOG",
-
-            Label               = {
-                text            = L["panel_mask_tips"],
-                fontObject      = GameFontWhiteTiny,
-                justifyH        = "CENTER",
-                justifyV        = "TOP",
-                location        = {
-                    Anchor("TOPLEFT", 8, -8),
-                    Anchor("BOTTOMRIGHT", -8, 8)
-                }
-            }
-        }
-    end
-
-    RECYCLE_MASKS.OnInit = RECYCLE_MASKS.OnInit + OnInit
-
     _SVDB = SVManager("AshToAsh_BlizzardSkin_DB", "AshToAsh_BlizzardSkin_CharDB")
     _SVDB:SetDefault{
         Templates                                                                                                           = {}