+++ Localization/zhCN.lua
@@ -6,10 +6,14 @@ if not L then return end
 
 L["menu_title"] = "暴雪皮肤"
 L["panel_mask_tips"] = "只有面板序号大于另一面板才会自动依附！"
+L["panel_moving_tips"] = "按住Alt键移动面板"
 L["tips"] = "提示"
 L["appearance"] = "外观"
 L["block_blizzard_unitframe"] = "屏蔽暴雪团队框架"
 L["block_blizzard_unitframe_tips"] = "不管怎样，同时运行ATA和暴雪团队框架总是没什么好处的。\n打开/关闭此选项将会重载界面"
+L["adjust_auto_attach_margin"] = "调整面板(%d)间距"
+L["adjust_auto_attach_margin_tips"] = "当面板自动依附后，你可以使用此选项调整间距。\n\n注意：这个选项并不会被保存！"
+L["adjust_auto_attach_margin_input_title"] = "输入间距值(-100~100)"
 
 L["visibility"] = "可见性"
 L["visibility_hide"] = "隐藏"